chore(webpack): move to dotfolder
